add deletePlan & calculate

+++ src/main/java/com/nerdnull/donlate/server/controller/PlanController.java
@@ -3,25 +3,25 @@ package com.nerdnull.donlate.server.controller;
 import com.nerdnull.donlate.server.controller.request.CreatePlanRequest;
 import com.nerdnull.donlate.server.controller.request.CreatePlanStateRequest;
 import com.nerdnull.donlate.server.controller.request.UpdatePlanRequest;
-import com.nerdnull.donlate.server.domain.PlanEntity;
 import com.nerdnull.donlate.server.domain.PlanStateEntity;
+import com.nerdnull.donlate.server.dto.CalculateParseDto;
 import com.nerdnull.donlate.server.dto.PlanDto;
 import com.nerdnull.donlate.server.dto.PlanStateDto;
-import com.nerdnull.donlate.server.repository.PlanRepository;
+import com.nerdnull.donlate.server.parse.CalculateParse;
 import com.nerdnull.donlate.server.service.PlanService;
 import com.nerdnull.donlate.server.service.PlanStateService;
 import lombok.extern.slf4j.Slf4j;
+import org.json.simple.JSONArray;
+import org.json.simple.JSONObject;
+import org.json.simple.parser.JSONParser;
+import org.json.simple.parser.ParseException;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import com.nerdnull.donlate.server.controller.response.PlanDetailResponse;
 import com.nerdnull.donlate.server.controller.response.Response;
-import com.nerdnull.donlate.server.dto.PlanDto;
-import com.nerdnull.donlate.server.service.PlanService;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
-import java.util.Optional;
+
+
+import java.util.List;
 
 @Slf4j
 @RequestMapping("/api/v1/plan")
@@ -30,13 +30,11 @@ public class PlanController {
 
     private final PlanService planService;
     private final PlanStateService planStateService;
-    private final PlanRepository planRepository;
 
     @Autowired
-    public PlanController(PlanService planService, PlanStateService planStateService, PlanRepository planRepository) {
+    public PlanController(PlanService planService, PlanStateService planStateService) {
         this.planService = planService;
         this.planStateService = planStateService;
-        this.planRepository = planRepository;
     }
 
     @PostMapping("/create")
@@ -59,7 +57,7 @@ public class PlanController {
         return Response.ok(1);
     }
 
-    @PostMapping("/update")
+    @PutMapping("/update")
     public Response<Integer> update(@RequestBody UpdatePlanRequest updatePlanRequest) {
         try {
             updatePlanRequest.isNotNull();
@@ -76,10 +74,22 @@ public class PlanController {
         return Response.ok(1);
     }
 
-    @PostMapping("/delete/{planId}")
-    public void delete(@PathVariable Long planId){
-
-        this.planService.deletePlan(planId);
+    @DeleteMapping(value = "/delete/{planId}")
+    public Response<String> delete(@PathVariable("planId") Long planId){
+        try {
+            if(planId==null) throw new IllegalArgumentException("planId could not be null");
+            this.planStateService.deleteByPlanId(planId);
+            this.planService.deletePlan(planId);
+        }
+        catch (IllegalArgumentException e){
+            log.error(e.getMessage(), e);
+            return Response.error(Response.BAD_REQUEST, e.getMessage());
+        }
+        catch (Exception e){
+            log.error(e.getMessage(), e);
+            return Response.error(Response.INTERNAL_SERVER_ERROR, e.getMessage());
+        }
+        return Response.ok("delete complete!!");
     }
 
     @PostMapping("/approve")
@@ -92,6 +102,24 @@ public class PlanController {
         catch (IllegalAccessException e){
             log.error(e.getMessage(), e);
             return Response.error(Response.BAD_REQUEST, e.getMessage());
+        }
+        catch (Exception e){
+            log.error(e.getMessage(), e);
+            return Response.error(Response.INTERNAL_SERVER_ERROR, e.getMessage());
+        }
+        return Response.ok(1);
+    }
+
+    @PutMapping("/calculate")
+    public Response<Integer>calculate(@RequestBody String body){
+        try{
+            CalculateParseDto cal = CalculateParse.parse(body);
+            PlanDto plan = this.planService.getDetails(cal.getPlanId());
+            List<PlanStateDto> planStateList = plan.getPlanStateList();
+            for (PlanStateDto p : planStateList) {
+                this.planStateService.setPlanState(new PlanStateDto(p.getPlanStateId(),cal.getPlanId(),
+                        p.getUserId(),null,null,cal.getUserState().get(p.getUserId())));
+            }
         }
         catch (Exception e){
             log.error(e.getMessage(), e);
@@ -113,5 +141,4 @@ public class PlanController {
             return Response.error(Response.INTERNAL_SERVER_ERROR, e.getMessage());
         }
     }
-
 }
